pyproject.toml
[tool.poetry]
name = "ticket-recognition-2024"
version = "0.1.0"
description = ""
authors = ["Eric CHI <ericchi@sjis.com.tw>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
efficientnet-pytorch = "^0.7.1"
GitPython = "^3.1.31"
matplotlib = "^3.7.1"
numpy = "^1.24.3"
opencv-python = "^4.7.0"
Pillow = "^9.5.0"
psutil = "^5.9.5"
PyYAML = "^6.0"
requests = "^2.30.0"
thop = "^0.1.1"
tqdm = "^4.65.0"
ultralytics = "^8.0.103"
pandas = "^2.0.1"
seaborn = "^0.12.2"
setuptools = "^67.7.2"
scikit-image = "^0.20.0"
pydantic = "^1.10.7"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-multipart = "^0.0.6"
loguru = "^0.7.0"

[tool.poetry.dev-dependencies]
yapf = "^0.33.0"
toml = "^0.10.2"
jupyterlab = "^3.6.3"
matplotlib = "^3.7.1"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
